tests/wpt/tests/content-security-policy/gen/top.http-rp/worker-src-none/sharedworker-module.http.html

<!DOCTYPE html>
<!-- DO NOT EDIT! Generated by `common/security-features/tools/generate.py --spec content-security-policy/` -->
<html>
  <head>
    <meta charset="utf-8">
    <meta name="timeout" content="long">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/security-features/resources/common.sub.js"></script>
    <script src="../../../generic/test-case.sub.js"></script>
  </head>
  <body>
    <script>
      TestCase(
        [
          {
            "expectation": "blocked",
            "origin": "same-http",
            "redirection": "keep-origin",
            "source_context_list": [],
            "source_scheme": "http",
            "subresource": "sharedworker-module",
            "subresource_policy_deliveries": [],
            "test_description": "Content Security Policy: Expects blocked for sharedworker-module to same-http origin and keep-origin redirection from http context."
          },
          {
            "expectation": "blocked",
            "origin": "same-http",
            "redirection": "no-redirect",
            "source_context_list": [],
            "source_scheme": "http",
            "subresource": "sharedworker-module",
            "subresource_policy_deliveries": [],
            "test_description": "Content Security Policy: Expects blocked for sharedworker-module to same-http origin and no-redirect redirection from http context."
          }
        ],
        new SanityChecker()
      ).start();
    </script>
    <div id="log"></div>
  </body>
</html>
