LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/worker-src-none/sharedworker-module.http-expected.txt


PASS Content Security Policy: Expects blocked for sharedworker-module to same-http origin and keep-origin redirection from http context.
FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and keep-origin redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
PASS Content Security Policy: Expects blocked for sharedworker-module to same-http origin and no-redirect redirection from http context.
FAIL Content Security Policy: Expects blocked for sharedworker-module to same-http origin and no-redirect redirection from http context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
